+++ image/install.sh
@@ -8,7 +8,6 @@ set -x
 /build/pups.sh
 /build/utilities.sh
 
-if [[ "$ruby18" = 1 ]]; then /build/ruby1.8.sh; fi
 if [[ "$ruby19" = 1 ]]; then /build/ruby1.9.sh; fi
 if [[ "$ruby20" = 1 ]]; then /build/ruby2.0.sh; fi
 if [[ "$ruby21" = 1 ]]; then /build/ruby2.1.sh; fi